chore(Tests):Improve Test coverage
