+++ test/transactions.test.js
@@ -2,7 +2,6 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import { getMaxListeners } from 'cluster';
 import app from '../SERVER/index';
 
 chai.should();
